+++ .gitignore
@@ -1,2 +1,2 @@
-frontend-react/node_modules/
-backend-express/node_modules/
+frontend/node_modules/
+backend/node_modules/
