--- backend-express/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "backend-express",
-  "version": "1.0.0",
-  "description": "",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "nodemon server.js",
-    "client": "npm start --prefix ../frontend-react"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.18.1",
-    "mongoose": "^6.6.1"
-  }
-}
