--- frontend-react/src/components/button.component.jsx
@@ -1,9 +1,0 @@
-import "./button.styles.scss";
-
-
-const Button = ({text}) => {
-    return(
-        <button className="button"> { text } world</button>
-    );
-}
-export default Button;
