+++ frontend/src/App.js
@@ -1,0 +1,11 @@
+import Button from './components/button.component';
+
+function App() {
+  return (
+    <div className="App">
+      <Button text="Hello" />
+    </div>
+  );
+}
+
+export default App;
