+++ frontend/src/components/button.component.jsx
@@ -1,0 +1,9 @@
+import "./button.styles.scss";
+
+
+const Button = ({text}) => {
+    return(
+        <button className="button"> { text } world</button>
+    );
+}
+export default Button;
